added ldes fragment service
