+++ ldes-server-simulator/index.ts
@@ -1,6 +1,7 @@
 import fastify from 'fastify'
 import minimist from 'minimist'
 import { LdesFragmentRepository } from './ldes-fragment-repository';
+import { LdesFragmentService } from './ldes-fragment-service';
 import { TreeNode } from './tree-specification';
 
 const server = fastify();
@@ -8,7 +9,8 @@ const args = minimist(process.argv.slice(2));
 console.debug("arguments: ", args);
 
 const baseUrl = new URL(args.baseUrl || 'http://localhost:8080')
-const repository = new LdesFragmentRepository(baseUrl);
+const repository = new LdesFragmentRepository();
+const service = new LdesFragmentService(baseUrl, repository);
 const redirections: any = {};
 
 server.get('/', async (_request, _reply) => {
@@ -29,7 +31,7 @@ server.get('/*', async (request, reply) => {
 
 server.post('/ldes', async (request, reply) => {
   reply.statusCode = 201;
-  return { path: repository.store(request.body as TreeNode) };
+  return { path: service.store(request.body as TreeNode) };
 });
 
 interface Redirection {
@@ -53,7 +55,7 @@ server.post('/redirect', async (request, _reply) => {
 const options = { port: Number.parseInt(baseUrl.port), host: baseUrl.hostname };
 server.listen(options, async (err, address) => {
   if (args.seed) {
-    await repository.seed(args.seed);
+    await service.seed(args.seed);
   }
   if (err) {
     console.error(err)
